Make Magic Wand work and bypass Haiku bitmap drawing bug
This patch makes the Magic Wand selector tool work properly, and fixes other glitches in selection caused by a bug in Haiku's 'FillPolygon'. It doesn't quite fill... it is one pixel short on the right, so a StrokePolygon is needed to fix things up. (I believe this is another result of Haiku's mandatory anti-aliasing. This does seem to have other effects on selection, too, that I can't fix, but they don't seem to be serious in practice.)

Fixes issue #68 .