ArtPaint: set size of setting based on type, not value
Setting the size of the data based on the value seems dangerous since
it's a void* - a bad value could cause the wrong size data.  Using the
size of the typeCode is safer. In my experience this caused the color
selection to not function, as the size of the data didn't match
sizeof(rgb_color).