ArtPaint: fixed bad sliders; updated tool setup dialogs for consistency
Fixed all of the broken sliders from last commit. Refactored to add a method to build
grid layouts for the number sliders. Added constants to store the inset and spacing
values. Updated all of the tool setup dialogs to have consistent layouts. Added new string
"Options" to replace some of the redundant uses of "Mode" in these dialogs.