ArtPaint: Fixed copy/paste when area is selected
When importing data into a bitmap, it MUST be the same size as the
source bitmap otherwise it doesn't work, which is why pasting a
selection into a layer didn't do anything. There's an alternate form
of ImportBits that allows a data size and offset, so I used that and
passed the offset of the selection through.

Fixes #121