Manipulators: refactor so selection is a class member
Refactored so that each GUI manipulator has a class member called
"selection" that holds the current selection when the manipulator is
instantiated. This is instead of passing in the selection only when
the bitmap is actually manipulated, because the selection should never
change during a manipulation. It allows other functions in the manipulator
to access the selection, for example if we wanted to draw a transform box
around the transform region, or set the rotation pivot to the center of
the selection.