Rename canvas "Scale..." to "Resize...", update en.catkeys
Now that the canvas menu items have no shortcuts, there's no
mnemonic need to share the menu label "Scale" and shortcut 'E'
between Edit and Canvas menu.

To avoid confusion (e.g. in the docs or when discussing among
users) it's better to go with another label for the canvas
scaling: "Resize"